+++ lib/hub/context.rb
@@ -220,6 +220,7 @@ module Hub
 
       def initialize(*args)
         super
+        self.name = self.name.tr(' ', '-')
         self.host ||= (local_repo || LocalRepo).default_host
         self.host = host.sub(/^ssh\./i, '') if 'ssh.github.com' == host.downcase
       end
@@ -432,7 +433,10 @@ module Hub
       editor = git_command 'var GIT_EDITOR'
       editor = ENV[$1] if editor =~ /^\$(\w+)$/
       editor = File.expand_path editor if (editor =~ /^[~.]/ or editor.index('/')) and editor !~ /["']/
-      editor.shellsplit
+      # avoid shellsplitting "C:\Program Files"
+      if File.exist? editor then [editor]
+      else editor.shellsplit
+      end
     end
 
     module System
@@ -441,7 +445,7 @@ module Hub
       # Returns an array, e.g.: ['open']
       def browser_launcher
         browser = ENV['BROWSER'] || (
-          osx? ? 'open' : windows? ? 'start' :
+          osx? ? 'open' : windows? ? %w[cmd /c start] :
           %w[xdg-open cygstart x-www-browser firefox opera mozilla netscape].find { |comm| which comm }
         )
 
@@ -480,6 +484,10 @@ module Hub
       # Returns a Boolean.
       def command?(name)
         !which(name).nil?
+      end
+
+      def tmp_dir
+        ENV['TMPDIR'] || ENV['TEMP'] || '/tmp'
       end
     end