+++ lib/hub/github_api.rb
@@ -66,7 +66,7 @@ module Hub
 
     # Public: Create a new project.
     def create_repo project, options = {}
-      is_org = project.owner != config.username(api_host(project.host))
+      is_org = project.owner.downcase != config.username(api_host(project.host)).downcase
       params = { :name => project.name, :private => !!options[:private] }
       params[:description] = options[:description] if options[:description]
       params[:homepage]    = options[:homepage]    if options[:homepage]
@@ -77,6 +77,7 @@ module Hub
         res = post "https://%s/user/repos" % api_host(project.host), params
       end
       res.error! unless res.success?
+      res.data
     end
 
     # Public: Fetch info about a pull request.
@@ -109,29 +110,21 @@ module Hub
       res.data
     end
 
-    # Return the pull request corresponding to the current branch
-    def get_pullrequest project, branch_name
-      for state in ['open', 'closed']
-        page = 1
-        res = nil
-        while page == 1 or res.data.length > 0
-          res = get "https://%s/repos/%s/%s/pulls?state=%s&page=%s" %
-            [api_host(project.host), project.owner, project.name, state, page]
-          res.error! unless res.success?
-          res.data.each { |x|
-            if branch_name == x['head']['label'].split(':', 0)[1]
-              return x['html_url']
-            end
-          }
-          page += 1
-        end
-      end
-      nil
+    def statuses project, sha
+      res = get "https://%s/repos/%s/%s/statuses/%s" %
+        [api_host(project.host), project.owner, project.name, sha]
+
+      res.error! unless res.success?
+      res.data
     end
 
     # Methods for performing HTTP requests
     #
-    # Requires access to a `config` object that implements `proxy_uri(with_ssl)`
+    # Requires access to a `config` object that implements:
+    # - proxy_uri(with_ssl)
+    # - username(host)
+    # - update_username(host, old_username, new_username)
+    # - password(host, user)
     module HttpMethods
       # Decorator for Net::HTTPResponse
       module ResponseMethods
@@ -145,7 +138,12 @@ module Hub
           data['errors'].map do |err|
             case err['code']
             when 'custom'        then err['message']
-            when 'missing_field' then "field '%s' is missing" % err['field']
+            when 'missing_field'
+              %(Missing field: "%s") % err['field']
+            when 'invalid'
+              %(Invalid value for "%s": "%s") % [ err['field'], err['value'] ]
+            when 'unauthorized'
+              %(Not allowed to change field "%s") % err['field']
             end
           end.compact if data['errors']
         end
@@ -159,10 +157,17 @@ module Hub
         perform_request url, :Post do |req|
           if params
             req.body = JSON.dump params
-            req['Content-Type'] = 'application/json'
+            req['Content-Type'] = 'application/json;charset=utf-8'
           end
           yield req if block_given?
-          req['Content-Length'] = req.body ? req.body.length : 0
+          req['Content-Length'] = byte_size req.body
+        end
+      end
+
+      def byte_size str
+        if    str.respond_to? :bytesize then str.bytesize
+        elsif str.respond_to? :length   then str.length
+        else  0
         end
       end
 
@@ -180,13 +185,17 @@ module Hub
           create_connection host_url
         end
 
+        req['User-Agent'] = "Hub #{Hub::VERSION}"
         apply_authentication(req, url)
         yield req if block_given?
-        res = http.start { http.request(req) }
-        res.extend ResponseMethods
-        res
-      rescue SocketError => err
-        raise Context::FatalError, "error with #{type.to_s.upcase} #{url} (#{err.message})"
+
+        begin
+          res = http.start { http.request(req) }
+          res.extend ResponseMethods
+          return res
+        rescue SocketError => err
+          raise Context::FatalError, "error with #{type.to_s.upcase} #{url} (#{err.message})"
+        end
       end
 
       def request_uri url
@@ -240,10 +249,18 @@ module Hub
         if (req.path =~ /\/authorizations$/)
           super
         else
+          refresh = false
           user = url.user || config.username(url.host)
           token = config.oauth_token(url.host, user) {
+            refresh = true
             obtain_oauth_token url.host, user
           }
+          if refresh
+            # get current user info user to persist correctly capitalized login name
+            res = get "https://#{url.host}/user"
+            res.error! unless res.success?
+            config.update_username(url.host, user, res.data['login'])
+          end
           req['Authorization'] = "token #{token}"
         end
       end
@@ -336,12 +353,19 @@ module Hub
       end
 
       def username host
+        return ENV['GITHUB_USER'] unless ENV['GITHUB_USER'].to_s.empty?
         host = normalize_host host
         @data.fetch_user host do
           if block_given? then yield
           else prompt "#{host} username"
           end
         end
+      end
+
+      def update_username host, old_username, new_username
+        entry = @data.entry_for_user(normalize_host(host), old_username)
+        entry['user'] = new_username
+        @data.save
       end
 
       def api_token host, user
@@ -354,6 +378,7 @@ module Hub
       end
 
       def password host, user
+        return ENV['GITHUB_PASSWORD'] unless ENV['GITHUB_PASSWORD'].to_s.empty?
         host = normalize_host host
         @password_cache["#{user}@#{host}"] ||= prompt_password host, user
       end
@@ -380,12 +405,14 @@ module Hub
         end
       end
 
-      # FIXME: probably not cross-platform
+      NULL = defined?(File::NULL) ? File::NULL :
+               File.exist?('/dev/null') ? '/dev/null' : 'NUL'
+
       def askpass
-        tty_state = `stty -g`
+        tty_state = `stty -g 2>#{NULL}`
         system 'stty raw -echo -icanon isig' if $?.success?
         pass = ''
-        while char = $stdin.getbyte and !(char == 13 or char == 10)
+        while char = getbyte($stdin) and !(char == 13 or char == 10)
           if char == 127 or char == 8
             pass[-1,1] = '' unless pass.empty?
           else
@@ -395,6 +422,15 @@ module Hub
         pass
       ensure
         system "stty #{tty_state}" unless tty_state.empty?
+      end
+
+      def getbyte(io)
+        if io.respond_to?(:getbyte)
+          io.getbyte
+        else
+          # In Ruby <= 1.8.6, getc behaved the same
+          io.getc
+        end
       end
 
       def proxy_uri(with_ssl)